[FP-37] Refactor: DateInput 모듈화
