+++ src/components/Input/index.ts
@@ -1,6 +1,6 @@
-import DateInput from './DateInput';
+import DateInput from './DateInput/dateInput';
 import DefaultInput from './DefaultInput';
-import DropdownInput from './DropdownInput/DropdownInput';
+import DropdownInput from './DropdownInput/dropdownInput';
 import PasswordInput from './PasswordInput';
 
 const Input = Object.assign(DefaultInput, {
